+++ backend/apid/apid.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/sensu/sensu-go/backend/authentication"
 
+	"github.com/coreos/etcd/client"
 	"github.com/coreos/etcd/clientv3"
 	"github.com/gorilla/mux"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
@@ -41,6 +42,7 @@ type APId struct {
 	tls                 *types.TLSOptions
 	cluster             clientv3.Cluster
 	etcdClientTLSConfig *tls.Config
+	client              client.Client
 }
 
 // Option is a functional option.
@@ -57,6 +59,7 @@ type Config struct {
 	Cluster             clientv3.Cluster
 	EtcdClientTLSConfig *tls.Config
 	Authenticator       *authentication.Authenticator
+	Client              client.Client
 }
 
 // New creates a new APId.
@@ -73,6 +76,7 @@ func New(c Config, opts ...Option) (*APId, error) {
 		cluster:             c.Cluster,
 		etcdClientTLSConfig: c.EtcdClientTLSConfig,
 		Authenticator:       c.Authenticator,
+		client:              c.Client,
 	}
 
 	// prepare TLS configs (both server and client)
@@ -101,7 +105,7 @@ func New(c Config, opts ...Option) (*APId, error) {
 	router := mux.NewRouter().UseEncodedPath()
 	router.NotFoundHandler = middlewares.SimpleLogger{}.Then(http.HandlerFunc(notFoundHandler))
 	router.Handle("/metrics", promhttp.Handler())
-	registerUnauthenticatedResources(router, a.store, a.cluster, a.etcdClientTLSConfig)
+	registerUnauthenticatedResources(router, a.store, a.cluster, a.etcdClientTLSConfig, a.client)
 	a.registerGraphQLService(router, c.URL, tlsClientConfig)
 	registerAuthenticationResources(router, a.store, a.Authenticator)
 	a.registerRestrictedResources(router)
@@ -186,6 +190,7 @@ func registerUnauthenticatedResources(
 	store store.Store,
 	cluster clientv3.Cluster,
 	etcdClientTLSConfig *tls.Config,
+	client client.Client,
 ) {
 	mountRouters(
 		NewSubrouter(
@@ -194,6 +199,7 @@ func registerUnauthenticatedResources(
 			middlewares.LimitRequest{},
 		),
 		routers.NewHealthRouter(actions.NewHealthController(store, cluster, etcdClientTLSConfig)),
+		routers.NewVersionRouter(actions.NewVersionController(store, client)),
 	)
 }