+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added per resource counts to tessen data collection.
 - Added event processing counts to tessen data collection.
 - Added ability to query for `Handlers` (individual and collections) from the GraphQL query endpoint.
+- Added `/version` to retrieve the current etcd server/cluster version and the sensu-backend version.
 
 ### Fixed
 - Only bury switchsets of checks that no longer have a TTL, in order to reduce